+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/stock/logic/service/ItemStockService.java
@@ -144,22 +144,26 @@ public abstract class ItemStockService<T extends ItemStock> {
         List<String> errors = new ArrayList<>();
         try {
             if (itemStock.getPartner() == null) {
-                errors.add("Missing Partner");
+                errors.add("Missing Partner.");
             }
             if (itemStock.getMaterial() == null) {
-                errors.add("Missing Material");
+                errors.add("Missing Material.");
             }
             if (itemStock.getLocationBpna() == null) {
-                errors.add("Missing locationBpna");
+                errors.add("Missing locationBpna.");
             }
             if (itemStock.getLocationBpns() == null) {
-                errors.add("Missing locationBpns");
+                errors.add("Missing locationBpns.");
             }
             if (itemStock.getMeasurementUnit() == null) {
-                errors.add("Missing measurementUnit");
+                errors.add("Missing measurementUnit.");
             }
             if (itemStock.getLastUpdatedOnDateTime() == null) {
-                errors.add("Missing lastUpdatedOnTime");
+                errors.add("Missing lastUpdatedOnTime.");
+            }
+            if (!((itemStock.getCustomerOrderId() != null && itemStock.getCustomerOrderPositionId() != null) || 
+                (itemStock.getCustomerOrderId() == null && itemStock.getCustomerOrderPositionId() == null && itemStock.getSupplierOrderId() == null))) {
+                errors.add("If an order position reference is given, customer order number and customer order position number must be set.");
             }
         } catch (Exception e) {
             log.error("Basic Validation failed: " + itemStock + "\n" + e.getMessage());
@@ -183,13 +187,13 @@ public abstract class ItemStockService<T extends ItemStock> {
             Material material = itemStock.getMaterial();
             MaterialPartnerRelation relation = mprService.find(material, partner);
             if (relation == null) {
-                errors.add("Missing MaterialPartnerRelation");
+                errors.add("Missing MaterialPartnerRelation.");
             }
             if (!material.isMaterialFlag()) {
-                errors.add("Material flag is missing");
+                errors.add("Material flag is missing.");
             }
             if (relation != null && !relation.isPartnerSuppliesMaterial()) {
-                errors.add("Partner does not supply material");
+                errors.add("Partner does not supply material.");
             }
         } catch (Exception e) {
             log.error("MaterialItemStock Validation failed: " + itemStock + "\n" + e.getMessage());
@@ -205,13 +209,13 @@ public abstract class ItemStockService<T extends ItemStock> {
             Material material = itemStock.getMaterial();
             MaterialPartnerRelation relation = mprService.find(material, partner);
             if (relation == null) {
-                errors.add("Missing MaterialPartnerRelation");
+                errors.add("Missing MaterialPartnerRelation.");
             }
             if (!material.isProductFlag()) {
-                errors.add("Product flag is missing");
+                errors.add("Product flag is missing.");
             }
             if (relation != null && !relation.isPartnerBuysMaterial()) {
-                errors.add("Partner does not buy material");
+                errors.add("Partner does not buy material.");
             }
         } catch (Exception e) {
             log.error("ProductItemStock Validation failed: " + itemStock + "\n" + e.getMessage());
@@ -227,13 +231,13 @@ public abstract class ItemStockService<T extends ItemStock> {
                     .filter(site -> site.getBpns().equals(itemStock.getLocationBpns()))
                     .findFirst().orElse(null);
             if (stockSite == null) {
-                errors.add("Site not found");
+                errors.add("Site not found.");
             }
             var stockAddress = stockSite != null ? stockSite.getAddresses().stream()
                     .filter(addr -> addr.getBpna().equals(itemStock.getLocationBpna()))
                     .findFirst().orElse(null) : null;
             if (stockAddress == null) {
-                errors.add("Address not found");
+                errors.add("Address not found for partner and site.");
             }
         } catch (Exception e) {
             log.error("Location Validation failed: " + itemStock + "\n" + e.getMessage());